Add bin/upload adapted from official sample script
I've hardly customized it for my own usage yet (except for adding a few
fields and selecting default values to my liking), nor added a progress
bar which is pretty important when uploading ~5GB videos which could
take a couple of minutes. But it works, here's a sample command line I
used earlier today:

    upload -t '20170323 SNH48 Team X 梦想的旗帜 21 沈阳巡演' \
           -d $'http://live.shy48.com/Index/invedio/id/32\n\n本视频版权归上海丝芭文化传媒集团有限公司所有。' \
           --tags 'SNH48,梦想的旗帜' \
           '20170323 Team X 梦想的旗帜 21 沈阳巡演.m2ts'

Note that a .ts file would be recognized as text/vnd.trolltech.linguist,
so need to hard link it to a .m2ts to have it uploaded properly. Which
is rather lame. Might be able to fix it (look for a MIME arg in the
API?), but not sure.