Add helper bin/download
The heart and soul of this channel.

This is a complete rewirte. For the past four+ months I've been using a
shell script with hard-coded variables that I change for each video. Now
I've morphed it into this Python script that is better modularized, with
more consistent logging, and external config files (which means two
different jobs can be run at once without duplicating and modifying the
script), albeit a fair bit more verbose.