bin/upload and bin/download: allow inserting into playlists...
when uploading.

data/playlists.json still needs to be populated, possibly with an
additional helper, for this to work (upload takes human-readable and
memorable playlist names in the --playlists argument; download reads
playlist names from config files).

Since I wrote these entirely offline, there might be bugs to be fixed in
addition to the data file.